build: introduce build configure script
Introduce a formal build configure script so that we can extend libtpa
more easily.

For example, we are considering to add AF_XDP pmd support, which requires
a new DPDK version. Without this script, we have to do two configs:

- enable AF_XDP
- set DPDK version to v22.11

Now kudos to this script, we only need to do one config, that is enabling
AF_XDP, then the script will fix the rest and make it right for you.

Signed-off-by: Yuanhan Liu <liuyuanhan.131@bytedance.com>