build: build all DPDK PMDs by default
To make the the libtpa.a as minimal as possible, libtpa builds DPDK
with only mlnx pmd enabled by default: that's the only NIC we support
so far.

That's going to be changed. We are planning to add more NIC support,
especially those virtual NICs, such as af_packet, af_xdp, vhost and
virti-user, etc. So that all can play libtpa, just with a linux system.
Not all have a Mellanox NIC after all.

Signed-off-by: Yuanhan Liu <liuyuanhan.131@bytedance.com>