arp: process ARP in datapath
With the help of flow bifurcation, libtpa doesn't have to handle ARP
reply in the datapath: it's handled in the ctrl path with the help
of raw socket.

For nics with no flow bifurcation support, the host TCP/IP stack no
longer exists, therefore, getting the reply by the raw socket won't
work any more. To make it work, we have to process ARP packets in
the datapath, including the ARP request and response packets.

Signed-off-by: Yuanhan Liu <liuyuanhan.131@bytedance.com>