shell: start the shell processing at the end of init
If one subsystem is not well initialized and the shell cmd is exposed to
user firstly, we may get a crash at the start time when user tries to
execute it.

Starting the shell processing at the end of init (when everything is well
initialized) would avoid things like that.

Signed-off-by: Yuanhan Liu <liuyuanhan.131@bytedance.com>